Move CalculatedColors back to AvatarTokens and handle logic for last Name only
